packaging: initial repo creation
oVirt Live is an unofficial spin of live fedora 17(currently),
based on oVirt 3.1 release using allInOne plugin.

This version can be used for POC or demos, and suppose to give a sneak
preview for the working product operation. simplify the the installation for
new users,and will allow users to experiment with oVirt before later
installing it.

Change-Id: I66cca8f5d5abaa710f054c59043ff6363b68256f
Signed-off-by: Moran Goldboim <mgoldboi@redhat.com>